truly fixed #44
